Remove native_member

+++ Bia/create_member.hpp
@@ -6,7 +6,8 @@
 
 #include "type_traits.hpp"
 #include "member.hpp"
-#include "native_member.hpp"
+#include "int_member.hpp"
+#include "double_member.hpp"
 #include "cstring_member.hpp"
 #include "static_function.hpp"
 #include "member_function.hpp"
@@ -55,7 +56,7 @@ template<typename Type>
 inline typename std::enable_if<std::is_integral<Type>::value>::type create_member(member * _destination, Type _value)
 {
 	if (_destination) {
-		_destination->replace_this<native::int_member>(static_cast<int64_t>(_value));
+		_destination->replace_this<native::int_member>(_value);
 	}
 }
